backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
docstring_parser==0.16
Faker==30.3.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
instructor==1.5.2
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
openai==1.51.2
openapi==2.0.0
propcache==0.2.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
referencing==0.35.1
rich==13.9.2
rpds-py==0.20.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tenacity==8.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
yarl==1.14.0
